+++ src/configuration/models/unite/uniteUnitTestEngine.ts
@@ -1,0 +1,6 @@
+/**
+ * Type for Unite Configuration unitEngine.
+ */
+export type UniteUnitTestEngine =
+    "PhantomJS" |
+    "ChromeHeadless";