+++ src/pipelineSteps/e2eTestRunner/protractor.ts
@@ -27,7 +27,7 @@ export class Protractor extends EnginePipelineStepBase {
     }
 
     public async process(logger: ILogger, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables): Promise<number> {
-        engineVariables.toggleDevDependency(["protractor", "browser-sync"], uniteConfiguration.e2eTestRunner === "Protractor");
+        engineVariables.toggleDevDependency(["protractor", "webdriver-manager", "browser-sync"], uniteConfiguration.e2eTestRunner === "Protractor");
 
         const esLintConfiguration = engineVariables.getConfiguration<EsLintConfiguration>("ESLint");
         if (esLintConfiguration) {
@@ -69,6 +69,7 @@ export class Protractor extends EnginePipelineStepBase {
         };
 
         defaultConfiguration.plugins = [];
+        defaultConfiguration.localSeleniumStandaloneOpts = { jvmArgs: []};
 
         this._configuration = ObjectHelper.merge(defaultConfiguration, this._configuration);
 
@@ -83,7 +84,28 @@ export class Protractor extends EnginePipelineStepBase {
     private createConfig(): string[] {
         let lines: string[] = [];
         lines = lines.concat(this._scriptStart);
+
+        lines.push("const fs = require('fs');");
+        lines.push("const path = require('path');");
+        lines.push("const webDriverPath = path.resolve('./node_modules/webdriver-manager/selenium/');");
         lines.push(`exports.config = ${JsonHelper.codify(this._configuration)};`);
+        lines.push("const files = fs.readdirSync((webDriverPath));");
+        lines.push("const jvmArgs = [];");
+        lines.push("files.forEach(file => {");
+        lines.push("    const lowerFile = file.toLowerCase();");
+        lines.push("    if (lowerFile.substr(-3) !== \"zip\" && lowerFile.substr(-6) !== \"tar.gz\" && lowerFile.substr(-3) !== \"xml\" && lowerFile.substr(-4) !== \"json\") {");
+        lines.push("        if (lowerFile.substr(0, 5) === \"gecko\") {");
+        lines.push("            jvmArgs.push('-Dwebdriver.gecko.driver=' + path.join(webDriverPath, file));");
+        lines.push("        } else if (lowerFile.substr(0, 6) === \"chrome\") {");
+        lines.push("            jvmArgs.push('-Dwebdriver.chrome.driver=' + path.join(webDriverPath, file));");
+        lines.push("        } else if (lowerFile.substr(0, 8) === \"iedriver\") {");
+        lines.push("            jvmArgs.push('-Dwebdriver.ie.driver=' + path.join(webDriverPath, file));");
+        lines.push("        } else if (lowerFile.substr(0, 18) === \"microsoftwebdriver\") {");
+        lines.push("            jvmArgs.push('-Dwebdriver.edge.driver=' + path.join(webDriverPath, file));");
+        lines.push("        }");
+        lines.push("    }");
+        lines.push("});");
+        lines.push("exports.config.localSeleniumStandaloneOpts.jvmArgs = jvmArgs;");
         lines = lines.concat(this._scriptEnd);
         lines.push(super.wrapGeneratedMarker("/* ", " */"));
         return lines;